requirements.txt
absl-py==2.3.1
asgiref==3.10.0
astunparse==1.6.3
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.4
click==8.3.0
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
daphne==4.0.0
deepface==0.0.92
Django==4.2.7
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==23.0.0
h5py==3.15.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.2
mtcnn==0.1.1
namex==0.1.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==10.4.0
protobuf==6.33.0
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.0.0
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
scipy==1.16.2
setuptools==70.0.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow_cpu==2.20.0
termcolor==3.1.0
tf_keras==2.20.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
zope.interface==8.0.1
